Dwitter 로그인 적용

+++ server/controller/tweet.js
@@ -19,17 +19,25 @@ export async function getTweetsById(req, res) {
 }
 
 export async function createTweet(req, res, next) {
-    const { text, name, username } = req.body;
-    const tweet = await tweetRepository.create(text, name, username);
+    const { text } = req.body;
+    const tweet = await tweetRepository.create(text, req.userId);
     res.status(201).json(tweet);
 }
 
 export async function updateTweet(req, res) {
     const id = req.params.id;
     const text = req.body.text;
-    const tweet = await tweetRepository.update(id, text);
+    const tweet = await tweetRepository.getById(id);
+    if(!tweet) {
+        return res.sendStatus(404);
+    }
+
+    if(tweet.userId !== req.userId) {
+        return res.sendStatus(403);
+    }
+    const updated = await tweetRepository.update(id, text);
     if(tweet) {
-        res.status(200).json(tweet);
+        res.status(200).json(updated);
     } else {
         res.status(404).json({ message: `Tweet Id(${id}) not Found`});
     }
@@ -37,6 +45,14 @@ export async function updateTweet(req, res) {
 
 export async function deleteTweet(req, res) {
     const id = req.params.id;
+    const tweet = await tweetRepository.getById(id);
+    if(!tweet) {
+        return res.sendStatus(404);
+    }
+
+    if(tweet.userId !== req.userId) {
+        return res.sendStatus(403);
+    }
     await tweetRepository.remove(id)
     res.sendStatus(204);
 }
